--- project3Group9/src/main/java/com/example/project3group9/api/controller/UserController.java
@@ -1,25 +1,0 @@
-package com.example.project3group9.api.controller;
-
-import com.example.project3group9.api.model.User;
-import com.example.project3group9.api.repositories.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-public class UserController {
-
-    private final UserRepository userRepository;
-
-    @Autowired
-    public UserController(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @GetMapping("/users")
-    public List<User> users(){
-        return userRepository.findAll();
-    }
-}
