--- project3Group9/src/test/java/com/example/project3group9/Project3Group9ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.project3group9;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Project3Group9ApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
