+++ src/main/java/com/example/project3group9/api/repositories/UserRepository.java
@@ -1,0 +1,9 @@
+package com.example.project3group9.api.repositories;
+
+import com.example.project3group9.api.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Integer> {
+}
