+++ src/main/resources/application.properties
@@ -1,0 +1,7 @@
+spring.application.name=project3Group9
+spring.datasource.url=jdbc:mariadb://project3-database.ctq0ys0sgn1s.us-west-1.rds.amazonaws.com:3306/innodb
+spring.datasource.username=admin
+spring.datasource.password=2CTk3cvaUNuJUVH4uLt6
+spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
+spring.jpa.hibernate.ddl-auto=none
+spring.jpa.show-sql=true
